refactored writeing and backup for --setup

+++ src/nbpages/nbsetup.py
@@ -55,22 +55,32 @@ def make_dir_if_needed(path):
         print(f"- {path} directory already exists")
     return
 
-def write_template_if_needed(dir, filename, content):
-    """Create template file if needed, and verify that it exists."""
-    path = os.path.join(dir, filename)
-    backup_path = ""
+def create_backup(path):
+    backup_path = None
     if os.path.isfile(path):
-        backup_path = path + datetime.datetime.now().strftime(".backup-%Y-%m-%d-%H-%M-%S")
+        backup_path = path + datetime.datetime.now().strftime(".backup-%Y-%m-%d-%H%M%S")
         print(f"- backing up {path} to {backup_path}")
         shutil.copy2(path, backup_path)
+    return backup_path
+
+def write_content(path, content):
     print(f"- writing {path}")
     with open(path, 'w') as file:
         file.write(content)
-    if backup_path:
+    return path
+
+def compare_and_remove_backup(path, backup_path):
+    if  os.path.isfile(path) and os.path.isfile(backup_path):
         if open(path).read() == open(backup_path).read():
             print(f"- {path} content unchanged, {backup_path} deleted.")
             os.remove(backup_path)
-    return
+
+def write_with_backup(directory, filename, content):
+    """Write file with backup if needed"""
+    path = os.path.join(directory, filename)
+    backup_path = create_backup(path)
+    write_content(path, content)
+    compare_and_remove_backup(path,  backup_path)
 
 def nbsetup(config_file="nbpages.cfg"):
     """Setup directories if needed with default configuration and templates."""
@@ -97,29 +107,23 @@ def nbsetup(config_file="nbpages.cfg"):
                 "data_subdir": "data",
                 }
 
-    config_file_backup = ""
-    if os.path.isfile(config_file):
-        config_file_backup = config_file + datetime.datetime.now().strftime(".backup-%Y-%m-%d-%H-%M-%S")
-        print(f"- backing up existing {config_file} to {config_file_backup}")
-        shutil.copy2(config_file, config_file_backup)
 
-    print(f"- writing {config_file}")
+    config_file_backup = create_backup(config_file)
+
     config = configparser.ConfigParser()
     config["nbpages"] = nbpages
     with open(config_file, "w") as f:
+        print(f"- writing {config_file}")
         config.write(f)
 
-    if config_file_backup:
-        if open(config_file).read() == open(config_file_backup).read():
-            print(f"- {config_file} unchanged, backup deleted.")
-            os.remove(config_file_backup)
+    compare_and_remove_backup(config_file, config_file_backup)
 
     # create directories if needed
     print(f"creating templates directory")
     make_dir_if_needed(nbpages["templates_dir"])
-    write_template_if_needed(nbpages["templates_dir"], 'notebook_header.tpl', notebook_header_tpl)
-    write_template_if_needed(nbpages["templates_dir"], 'index.md.tpl', index_md_tpl)
-    write_template_if_needed(nbpages["templates_dir"], 'notebook.tpl', notebook_tpl)
+    write_with_backup(nbpages["templates_dir"], 'notebook_header.tpl', notebook_header_tpl)
+    write_with_backup(nbpages["templates_dir"], 'index.md.tpl', index_md_tpl)
+    write_with_backup(nbpages["templates_dir"], 'notebook.tpl', notebook_tpl)
 
     print(f"creating source directory")
     make_dir_if_needed(nbpages["src_dir"])
